SpriteAnim working with test sprite sheet
[git-p4: depot-paths = "//depot/C35/Students/cwsullivan/SD/": change = 446081]

+++ Engine/Code/Engine/Renderer/BitmapFont.cpp
@@ -86,7 +86,7 @@ void BitmapFont::AddVertsForTextInBox2D(std::vector<Vertex>& verts, std::string 
 	Vec2 startPos = box.m_mins + offset;
 
 	int glyphsDrawn = 0;
-	int splitStringSize = splitStrings.size() - 1;
+	int splitStringSize = static_cast<int>( splitStrings.size() ) - 1;
 	for (int stringIndex = 0; stringIndex <= splitStringSize; ++stringIndex)
 	{
 		Vec2 linePos = startPos;
